refactor : mvc 모델로 패키지 분리
