+++ src/test/java/racing/CarTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import racing.domain.Car;
+
 class CarTest {
 
     @ParameterizedTest
